==> ImmuneCellAnalysis/04b_DeconvoluteBulk/04b_MultivariateCox_fromtemplate.Rmd <==
---
title: "Multivariate Cox regression analysis"
output: html_document
---

Run multivariate cox regression analysis based on predicted proportion of programs. Control for confounders like tumor location, age, sex

## Overall steps:

  1. Determine which clinical variables to control for- run univariate cox to identify variables that may have an effect on survival
  2. Determine baseline level of effect of gene set score on survival- univariate cox. Can compare to kaplan-meier survival results (should be very similar)
  3. Determine effect of gene set score on survival when clinical variables are included in model- multivariate cox
  
```{r}
library(survival)
library(survminer)
library(dplyr)

## Set working/script directory
base_dir = "/Users/jlabelle/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/04b_DeconvoluteBulk/")
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/"

## Load helper functions
helper_functions<- paste0(base_dir, script_dir,"Survival_HelperFunctions.R")
source(helper_functions)


## Set directories
data_dir<- paste0("analysis/bulk_deconvolute/")
analysis_dir<- paste0("analysis/coxRegression/")
figure_dir<- paste0("figures/coxRegression/")
if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)};if(!dir.exists(analysis_dir)){dir.create(analysis_dir)}

## Load colors
tcell_colors<- readRDS("../plot_colors/Tcell.programs_rank8sharedByunion_reAnnot.Rds")
myeloid_colors<- readRDS("../plot_colors/Myeloid_program_colors.Rds")
```

## Load in data
Input:
list of "FinalMeta" from SurvivalAnalysis_template.Rmd

Minimum columns required: 
  a. SampleID ("submitter_id") 
  b. time to death ("time") 
  c. alive/dead ("censored"; alive= F, dead=T). Should be set as logical. 
  d. At least 1 method for splitting samples. Can be 1) raw scores 2) high/low scores 3) raw proportion 4) high/low proportion. Each geneset/program should have its own column

2.Clinical metadata. Contains variables you want to control for- age/sex/diagnosis/treatment/etc. First column should be sample ID. 
Note: should remove any variables with NA values for any sample, or remove those samples.

## List all files, subset as needed. Set analysis name based on cohorts
```{r}
## Load in processed data (03_DeconvoluteBulke_Preprocess.Rmd)
files<- list.files(data_dir, pattern="HighLowPropSplit")

## Subset cohorts as needed
#files<- files[!(grepl("K27M", files)| grepl("WT", files))] ## CBTN.DFCI.TCGA
#files<- files[grep("CBTN_All", files)] ##  "onlyCBTN"
#files<- files[grep("DFCI", files)] ## "onlyDFCI"
#files<- files[grep("TCGA", files)] ## "onlyTCGA"
files<- files[grepl(paste(c("CBTN_All"), collapse = "|"), files)]
#files<- files[grepl(paste(c("EPN"), collapse = "|"), files)]

analysis_name<- "OnlyCBTN_noDNA"
```

## Reformat proportions/survival, flatten into single df for each major cell type (CD4, CD8, my)
```{r}
surv_obj_list_allCellTypes<- lapply(files, function(x){
  tmp<- readRDS(paste0(data_dir, x))
  return(tmp)
})
names(surv_obj_list_allCellTypes)<- gsub("HighLowPropSplit_clinical_", "", gsub(".Rds", "", files))

## Add cohort/cell type info to each
final_meta_list<- lapply(names(surv_obj_list_allCellTypes), function(cohort){
  surv_obj_allCellTypes<- surv_obj_list_allCellTypes[[cohort]]
  finalMeta_allCellTypes<- lapply(names(surv_obj_allCellTypes), function(cellType){
    finalMeta<- surv_obj_allCellTypes[[cellType]]
    finalMeta$cohort<- cohort
    finalMeta$cellType<- cellType
    return(finalMeta)
  })
  names(finalMeta_allCellTypes)<- names(surv_obj_allCellTypes)
  return(finalMeta_allCellTypes)
  })
names(final_meta_list)<-names(surv_obj_list_allCellTypes)


## Flatten into 1 finalMeta df for each celltype. Need to subset to shared columns between all cohorts (varying clinical data)
finalMeta_allCellTypes<- lapply(c("cd8", "cd4", "myeloid", "GOI"), function(cellType){
  tmp<- lapply(final_meta_list, function(x){x[[cellType]]})
  
  ## Get vector of columsn that occur in all- time/censored, programs, any shared clinical data
  shared_columns<- unlist(unname(lapply(tmp, function(x){colnames(x)})))
  shared_columns<- table(shared_columns)
  shared_columns<- names(shared_columns)[shared_columns==length(tmp)]
  
  ## Subset to shared columns
  tmp<- lapply(tmp, function(x){x[,colnames(x) %in% shared_columns]})
  
  tmp<- do.call("rbind", tmp)
  return(tmp)
})
names(finalMeta_allCellTypes)<- c("cd8", "cd4", "myeloid", "GOI")
finalMeta_allCellTypes<- finalMeta_allCellTypes[!unlist(lapply(finalMeta_allCellTypes, is.null))]

## If analysis name == OnlyCBTN_noDNA: remove samples that are "DataNotAvailable" for Histone mutation
if(analysis_name=="OnlyCBTN_noDNA"){
  finalMeta_allCellTypes<- lapply(finalMeta_allCellTypes, function(x){
    x[x$HistoneStatus != "DataNotAvailable",]
  })
}

## Set high/low to factor- so that comparison is always to "high" program proportion
genesets<-c("Memory", "Predysfunctional", "Cytotoxic_NKreceptor", "IFN_Response", "Stress", "Effector_Memory", "Cycling",
             "Treg", "Inhibitory",
            "IFN_TAM", "Monocytes", "Inflammatory", "SEPP1_Mo_TAM", "Hypoxic_TAM", "Stress_Response",
            "KLRB1", "PTGER2", "SELPLG", "GNLY", "PDCD1")
finalMeta_allCellTypes<- lapply(finalMeta_allCellTypes, function(x){
  tmp<- x[,colnames(x) %in% genesets]
  tmp<- tmp %>% mutate_if(is.character, factor, levels=c("Low","High")) %>% as.data.frame()
  tmp<- cbind(tmp, x[,!colnames(x) %in% genesets])
  return(tmp)

})
```


## Set variables to potentially control for + whether to run as high/low or numeric with actual proportions
## These are variables from your clinical metadata that you think may have an effect on survival.
## Can also just use all variables
```{r}
head(finalMeta_allCellTypes$cd8)

## List with variables to control for (if running across multiple cohorts, may want to include "Cohort")
#ToControlFor<- c( "Age", "Sex", "race", "project_id", "TumorType_detailed")
#ToControlFor<- c("Age", "Sex", "TumorType_detailed", "Primary_Recurrent", "HistoneStatus")
#ToControlFor<- c("Age", "Sex", "Primary_Recurrent", "HistoneStatus", "Location")
ToControlFor<- c("Age", "Sex", "Primary_Recurrent", "HistoneStatus", "TumorType_detailed")
ToControlFor<- c("Subtype", "Gender", "Location", "Chemo", "Radio", "Progression", "Age", "Re")
ToControlFor<- c("Primary_Recurrent", "TumorType_detailed", "HistoneStatus") ## Final variables controlled for

split_method<- "" ## "" for high/low (default), "_numericProp" for actual proportions
```


                      
## Run univariate cox to determine which variables to control for
## RunCoxWrapper function input: 
geneSet_groups = vector of genesets names to apply wrapper to
FinalMeta = "FinalMeta"- contains time, censored, and all genesets in geneSet_groups
controlFor = whether or not this will be univariate (FALSE) or multivariate (TRUE)
mySubset = which samples used in analysis- only used for saving analysis. useful when trying out multiple subsets
exportName = analysis name to use when saving data
exportCsv = whether or not to export csv file of results
subsetToHighLow = when splitting 25/50/25, subset to just 25/25

#########
## CD8 ##
#########
```{r}
## Run univariate cox for all potential controlling variables
## set exportCsv to FALSE if you want to examine results before exporting
allSamples_diagnosis<- RunCoxWrapper(geneSet_groups = ToControlFor, AllMeta= finalMeta_allCellTypes$cd8, mySubset="AllSamples", 
                                     controlFor=FALSE, 
                                     exportName = paste0("CD8_univariate_variables_", analysis_name), 
                                     exportCsv=FALSE, resultsFolder = analysis_dir)

## select just variables that had significant hazard
cox_df<-na.omit(allSamples_diagnosis$cox_df)
sig<-cox_df[cox_df$`Pr(>|z|)` < 0.1,]
sig_variables<-unique(sig$Variable)

ToBeControlled<- sig_variables
#ToBeControlled<- c(sig_variables, "HistoneStatus")
```


## Run univariate cox for genesets
## Univariate
```{r}
genesets_test<- c("Memory", "Predysfunctional", "Cytotoxic_NKreceptor", "IFN_Response", "Stress", "Effector_Memory", "Cycling")
if(split_method=="_numericProp"){
  genesets_test<- paste0(genesets_test, "_perCells")
}

## Run univariate for all samples
allSamples<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$cd8, mySubset=analysis_name, 
                                    controlFor=FALSE, exportName = paste0("CD8_univariate", split_method),
                                    exportCsv=TRUE, subsetToHighLow = TRUE, resultsFolder = analysis_dir,
                           appendToNumericVariable = "")
```

## Run multivariate cox, for all gene sets, controlling for variables

```{r}
## Control for all sig clinical confounders- can also specify only a subset of these
allSamples_diagnosis<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$cd8, mySubset=analysis_name,
                                     controlFor=ToBeControlled,
                                     exportName=paste0("CD8_ControlFor.", paste(ToBeControlled, collapse = "."), split_method), 
                                     exportCsv = TRUE, resultsFolder = analysis_dir)


```

## Testing proportionality assumption
## Check this for at least one of the above models
## Tests whether it's appropriate to apply regression model to these variables
## If p<0.05, may need to stratify by that variable rather than controlling for it
```{r}
lapply(allSamples_diagnosis$cox_objs, function(x){cox.zph(x)})
```


#########
## CD4 ##
#########
```{r}
## Run univariate cox for all potential controlling variables
## set exportCsv to FALSE if you want to examine results before exporting
allSamples_diagnosis<- RunCoxWrapper(geneSet_groups = ToControlFor, AllMeta= finalMeta_allCellTypes$cd4, mySubset=analysis_name, 
                                     controlFor=FALSE, 
                                     exportName = "CD4_univariate_variables", exportCsv=FALSE, resultsFolder = analysis_dir)

## select just variables that had significant hazard
cox_df<-na.omit(allSamples_diagnosis$cox_df)
sig<-cox_df[cox_df$`Pr(>|z|)` < 0.1,]
sig_variables<-unique(sig$Variable)

ToBeControlled<- sig_variables
#ToBeControlled<- "HistoneStatus"
```


## Run univariate cox for genesets
## Univariate
```{r}
genesets_test<- c("Memory", "Treg", "Inhibitory", "IFN_Response", "Stress", "Effector_Memory", "Cycling")
if(split_method=="_numericProp"){
  genesets_test<- paste0(genesets_test, "_perCells")
}

## Run univariate for all samples
allSamples<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$cd4, mySubset=analysis_name, 
                                    controlFor=FALSE, exportName = paste0("CD4_univariate", split_method),
                                    exportCsv=TRUE, subsetToHighLow = TRUE, resultsFolder = analysis_dir,
                           appendToNumericVariable = "")
```

## Run multivariate cox, for all gene sets, controlling for variables

```{r}
## Control for all sig clinical confounders- can also specify only a subset of these
allSamples_diagnosis<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$cd4, mySubset=analysis_name,
                                     controlFor=ToBeControlled,
                                     exportName=paste0("CD4_ControlFor.", paste(ToBeControlled, collapse = "."), split_method), 
                                     exportCsv = TRUE, resultsFolder = analysis_dir,
                                     appendToNumericVariable = "")


```



#############
## Myeloid ##
#############
```{r}
## Run univariate cox for all potential controlling variables
## set exportCsv to FALSE if you want to examine results before exporting
allSamples_diagnosis<- RunCoxWrapper(geneSet_groups = ToControlFor, AllMeta= finalMeta_allCellTypes$myeloid, mySubset=analysis_name, 
                                     controlFor=FALSE, 
                                     exportName = "Myeloid_univariate_variables", exportCsv=FALSE, resultsFolder = analysis_dir)

## select just variables that had significant hazard
cox_df<-na.omit(allSamples_diagnosis$cox_df)
sig<-cox_df[cox_df$`Pr(>|z|)` < 0.1,]
sig_variables<-unique(sig$Variable)

ToBeControlled<- sig_variables
#ToBeControlled<- "HistoneStatus"
```


## Run univariate cox for genesets
## Univariate
```{r}
genesets_test<-  c("IFN_TAM", "Monocytes", "Inflammatory", "SEPP1_Mo_TAM", "Hypoxic_TAM", "Stress_Response")
tmp<- apply(finalMeta_allCellTypes$myeloid[,genesets_test], 2, function(x){length(unique(x))})
genesets_test<- names(tmp)[tmp==2]

if(split_method=="_numericProp"){
  genesets_test<- paste0(genesets_test, "_perCells")
}


## Run univariate for all samples
allSamples<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$myeloid, mySubset=analysis_name, 
                                    controlFor=FALSE, exportName = paste0("Myeloid_univariate",split_method),
                                    exportCsv=TRUE, subsetToHighLow = TRUE, resultsFolder = analysis_dir,
                           appendToNumericVariable = "")
```

## Run multivariate cox, for all gene sets, controlling for variables

```{r}
## Control for all sig clinical confounders- can also specify only a subset of these
allSamples_diagnosis<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$myeloid, mySubset=analysis_name,
                                     controlFor=ToBeControlled,
                                     exportName=paste0("Myeloid_ControlFor.", paste(ToBeControlled, collapse = "."), split_method), 
                                     exportCsv = TRUE, resultsFolder = analysis_dir,
                                     appendToNumericVariable = "")


```

#########
## GOI ##
#########
```{r}
## Run univariate cox for all potential controlling variables
## set exportCsv to FALSE if you want to examine results before exporting
allSamples_diagnosis<- RunCoxWrapper(geneSet_groups = ToControlFor, AllMeta= finalMeta_allCellTypes$GOI, mySubset="AllSamples", 
                                     controlFor=FALSE, 
                                     exportName = paste0("GOI_univariate_variables_", analysis_name), 
                                     exportCsv=FALSE, resultsFolder = analysis_dir)

## select just variables that had significant hazard
cox_df<-na.omit(allSamples_diagnosis$cox_df)
sig<-cox_df[cox_df$`Pr(>|z|)` < 0.1,]
sig_variables<-unique(sig$Variable)

ToBeControlled<- sig_variables
#ToBeControlled<- c(sig_variables, "HistoneStatus")
```


## Run univariate cox for genesets
## Univariate
```{r}
genesets_test<- c("Memory", "Predysfunctional", "Cytotoxic_NKreceptor", "IFN_Response", "Stress", "Effector_Memory", "Cycling")
if(split_method=="_numericProp"){
  genesets_test<- paste0(genesets_test, "_perCells")
}

## Run univariate for all samples
allSamples<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$cd8, mySubset=analysis_name, 
                                    controlFor=FALSE, exportName = paste0("CD8_univariate", split_method),
                                    exportCsv=TRUE, subsetToHighLow = TRUE, resultsFolder = analysis_dir,
                           appendToNumericVariable = "")
```

## Run multivariate cox, for all gene sets, controlling for variables

```{r}
## Control for all sig clinical confounders- can also specify only a subset of these
allSamples_diagnosis<- RunCoxWrapper(genesets_test, finalMeta_allCellTypes$cd8, mySubset=analysis_name,
                                     controlFor=ToBeControlled,
                                     exportName=paste0("CD8_ControlFor.", paste(ToBeControlled, collapse = "."), split_method), 
                                     exportCsv = TRUE, resultsFolder = analysis_dir)


```


## Merge results together into a single df in order to summarize in a plot
Useful when you run multiple cohorts separately, but want to get an overview of all cohorts together
May need to modify for your data
```{r}
genesets_test<-  c("IFN_TAM", "Monocytes", "Inflammatory", "SEPP1_Mo_TAM", "Hypoxic_TAM", "Stress_Response",
                   "Memory", "Treg", "Inhibitory", "IFN_Response", "Stress", "Effector_Memory", "Cycling",
                   "Memory", "Predysfunctional", "Cytotoxic_NKreceptor", "IFN_Response", "Stress", "Effector_Memory", "Cycling")

## Get all results files
multi_files<-list.files(analysis_dir, pattern=paste0("ControlFor.", paste(ToBeControlled, collapse="."),"_", split_method))
multi_files<- multi_files[grepl(analysis_name, multi_files) & !grepl("allCoxResults", multi_files)]
uni_files<-list.files(analysis_dir, pattern=paste0("univariate", split_method))
uni_files<- uni_files[!grepl("variables_", uni_files)& grepl(analysis_name, uni_files)]

if(split_method==""){
  multi_files<- multi_files[!grepl("numericProp", multi_files)]
  uni_files<- uni_files[!grepl("numericProp", uni_files)]
}

## Initiate dataframe with univariate file:
  ## Read in first file
  AllCox<- read.csv(paste0(analysis_dir, uni_files[1]))
  
  #Subset to variables of interest, rename
  AllCox<-AllCox[,colnames(AllCox) %in% c("X", "exp.coef.", "Pr...z..")]
  colnames(AllCox)<- c("Variable", "HazardRatio", "pvalue")
  
  ## Add "ControlFor" column- "None" for univariate
  AllCox$ControlFor<- "None"
  
  ## Add major cell type
  AllCox$CellType<- gsub("_.*", "" ,uni_files[1])

## cycle through rest of univariate, adding to univariate df
  for (i in uni_files[2:length(uni_files)]){
    df<- read.csv(paste0(analysis_dir, i))
    df<-df[,colnames(df) %in% c("X", "exp.coef.", "Pr...z..")]
    colnames(df)<- c("Variable", "HazardRatio", "pvalue")
    
    ## Add "ControlFor" column- "None" for univariate
    df$ControlFor<-"None"
    
    ## Add major cell type
    df$CellType<- gsub("_.*", "" ,i)
    
    AllCox<- rbind(AllCox, df)
  }
  
  
## cycle through rest of cox (multivariate), adding to univariate df
  for (i in multi_files){
    df<- read.csv(paste0(analysis_dir, i))
    df<-df[,colnames(df) %in% c("X", "exp.coef.", "Pr...z..")]
    colnames(df)<- c("Variable", "HazardRatio", "pvalue")
    
    #Get name of variable being controlled for, add to dataframe
    ControlFor<- gsub("_results.csv", "", 
                 gsub(paste0("_", analysis_name), "",
                 gsub("CD4_", "", gsub("CD8_", "", gsub("Myeloid_", "", i )))))
    df$ControlFor<-ControlFor
    
    ## Remove any rows that are the variable being controlled for- not of interest
    df2<- df[grepl(paste(genesets_test, collapse = "|"),df$Variable ),]

    
    ## Add major cell type
    df2$CellType<- gsub("_.*", "" ,i)
    
    AllCox<- rbind(AllCox, df2)
  }


write.csv(AllCox,paste0(analysis_dir, "allCoxResults_ControlFor.", 
                        paste(ToBeControlled, collapse="."),
                        "_", analysis_name,
                        split_method,
                        ".csv"))
```

## Plot results
```{r}
analysis_name<- "OnlyCBTN_noDNA"
ToBeControlled<- c("Primary_Recurrent", "TumorType_detailed", "HistoneStatus")

## Read in merged results
df<- read.csv(paste0(analysis_dir, "allCoxResults_ControlFor.", paste(ToBeControlled, collapse="."),"_", analysis_name,split_method, ".csv"))

## Set sig values for alpha
df$tmp<-df$pvalue<0.05
df$tmp2<- df$pvalue<0.1
df$Sig<-rowSums(df[,c("tmp", "tmp2")])
df$Sig<-gsub("0", "NotSig", gsub("1", "AlmostSig", gsub("2", "Sig", df$Sig)))
df$Sig<-factor(df$Sig, levels=c("NotSig", "AlmostSig", "Sig"))
df$Variable<- gsub("\\.", "_", df$Variable)

## Rename colors as needed
colors_use<- c(tcell_colors, myeloid_colors$rank6)
if(split_method=="_numericProp"){
  names(colors_use)<- paste0(names(colors_use), "_perCells")
}else{
   names(colors_use)<- paste0(names(colors_use), "High")
}
names(colors_use)<- gsub("\\.", "_", names(colors_use))

## Variable to facet wrap by
df$Effect<- ifelse(df$HazardRatio>1, "Worse", "Better")

## Plot
ggplot(df, aes(x=HazardRatio, y=ControlFor,color=Variable ))+
  geom_jitter(size=8, width=0.01, height=0.2, aes(alpha=Sig))+
  geom_vline(xintercept=1)+
  theme_bw()+
  xlab("Hazard Ratio")+
  theme(axis.title.y = element_blank(),
        axis.text.y = element_text(face="bold", color="black", size="12"),
        legend.position = "bottom")+
  facet_grid(rows=vars(CellType), space="free_y", cols=vars(Effect), scale="free")+
  #facet_grid()+
  scale_color_manual(values=colors_use)+
  scale_alpha_manual(aes(alpha=Sig), values = c(NotSig=.2, AlmostSig=.5, Sig=1), guide="none")
ggsave(paste0(figure_dir, "allCoxResults_ControlFor.", 
              paste(ToBeControlled, collapse="."),"_", 
              analysis_name,split_method, ".pdf"), device="pdf", width=13, height=5)

## RAW DATA EXPORT
write.csv(df, file="../../notUploaded/ImmuneDraft_Rebecca/raw_results_files/cox_regression_survival_by_program.csv")
```


## Merge results across multiple analyses:
1) Different cohort combinations (single cohort, merged)- univariate + control for all sig confounders for each
2) Split by high/low + numeric proportions
```{r}
## Get all merged cox result files
files<- list.files(analysis_dir, pattern="allCoxResults_ControlFor")


## Read in + add on cohort, controlled for, numeric or highLow
all_res<- lapply(files, function(x){
  tmp<- read.csv(paste0(analysis_dir, x))
  
  ## Extract cohort info
  cohort<- ifelse(grepl("numericProp", x),
                  gsub(pattern = "(.*_)(.*)(_numericProp.*)",replacement = "\\2",x = x),
                  gsub(pattern = "(.*_)(.*)(\\.csv.*)",replacement = "\\2",x = x))
  tmp$cohort<- cohort
  
  ## Add on numeric or highLow
  tmp$split_method<- ifelse(grepl("numericProp", x), "numericProp", "highLow")
  return(tmp)
})
all_res_df<- do.call("rbind", all_res)

## For numeric proportion analysis, hazard ratio can be at much higher scale. set max hazard ratio
max_hr<- max(all_res_df$HazardRatio[all_res_df$split_method=="highLow"]) + 1
all_res_df$HazardRatio_thresh<- ifelse(all_res_df$HazardRatio>max_hr, max_hr, all_res_df$HazardRatio)

## Remove "perCells" and "High" from variable
all_res_df$Variable<- gsub("_perCells", "", gsub("High", "", all_res_df$Variable))
names(colors_use)<- gsub("_perCells", "", gsub("High", "", names(colors_use)))

## Variable to facet wrap by
all_res_df$Effect<- ifelse(all_res_df$HazardRatio>1, "Worse", "Better")

## Set sig values for alpha
all_res_df$tmp<-all_res_df$pvalue<0.05
all_res_df$tmp2<- all_res_df$pvalue<0.1
all_res_df$Sig<-rowSums(all_res_df[,c("tmp", "tmp2")])
all_res_df$Sig<-gsub("0", "NotSig", gsub("1", "AlmostSig", gsub("2", "Sig", all_res_df$Sig)))
all_res_df$Sig<-factor(all_res_df$Sig, levels=c("NotSig", "AlmostSig", "Sig"))
all_res_df$Variable<- gsub("\\.", "_", all_res_df$Variable)

## Merge control-for info + split (y axis)
all_res_df$yaxis<- paste0(all_res_df$split_method, "_", all_res_df$ControlFor)

## Split by major cell type- will plot separately
all_res_split<- split(all_res_df, f=all_res_df$CellType)

## Plot it all somehow
all_plots<- lapply(all_res_split, function(x){
  ggplot(x, aes(x=HazardRatio_thresh, y=yaxis,color=Variable ,alpha=Sig))+
    geom_jitter(size=8, width=0.01, height=0.2)+
    geom_vline(xintercept=1)+
    theme_bw()+
    xlab("Hazard Ratio")+
    theme(axis.title.y = element_blank(),
          axis.text.y = element_text(face="bold", color="black", size=12),
          plot.title = element_text(face="bold", color="black", size=16))+
    facet_grid(rows=vars(cohort), space="free_y", cols=vars(Effect), scale="free")+
    #facet_grid()+
    scale_color_manual(values=colors_use[names(colors_use) %in% x$Variable])+
    ggtitle(unique(x$CellType))
})
cowplot::plot_grid(plotlist = all_plots, ncol=3)
ggsave(paste0(figure_dir, "CoxRes_AllCohorts.AllSplits_062822.png"), width=40, height=8)
```
